+++ matholymp/roundupreg/actions.py
@@ -73,7 +73,7 @@ from matholymp.roundupreg.roundupsource import RoundupDataSource
 from matholymp.roundupreg.rounduputil import scores_from_str, \
     person_is_contestant, contestant_code, scores_final, \
     valid_country_problem, valid_int_str, create_rss, country_from_code, \
-    person_is_remote
+    person_is_remote, show_scores
 from matholymp.roundupreg.userauditor import valid_address
 
 
@@ -248,6 +248,8 @@ class ScoresCSVAction(Action):
             raise ValueError('This action only applies to people')
         if self.nodeid is not None:
             raise ValueError('Node id specified for CSV generation')
+        if not show_scores(self.db, self.db.getuid()):
+            raise Unauthorised('Scores are currently hidden')
         self.client.setHeader('Content-Type', 'text/csv; charset=UTF-8')
         self.client.setHeader('Content-Disposition',
                               'attachment; filename=scores.csv')
@@ -268,7 +270,8 @@ class PeopleCSVAction(Action):
         self.client.setHeader('Content-Disposition',
                               'attachment; filename=people.csv')
         show_all = self.hasPermission('Omnivident')
-        return RoundupSiteGenerator(self.db).people_csv_bytes(show_all)
+        return RoundupSiteGenerator(self.db).people_csv_bytes(
+            show_all, show_scores(self.db, self.db.getuid()))
 
 
 class MedalBoundariesCSVAction(Action):
@@ -428,6 +431,8 @@ class ScoresRSSAction(Action):
         """Output the RSS feed for scores."""
         if self.classname != 'country':
             raise ValueError('This action only applies to countries')
+        if not show_scores(self.db, self.db.getuid()):
+            raise Unauthorised('Scores are currently hidden')
 
         self.client.setHeader('Content-Type', 'application/rss+xml')